% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBands.R
\name{confidenceBands}
\alias{confidenceBands}
\title{Confidence Bands for Wasserstein Regression}
\usage{
confidenceBands(
  wass_regress_res,
  Xpred_df,
  level = 0.95,
  delta = 0.01,
  type = "density",
  figure = TRUE,
  fig_num = NULL
)
}
\arguments{
\item{wass_regress_res}{an object returned by the \code{wass_regress} function}

\item{Xpred_df}{k-by-p matrix (or dataframe, or named vector) used for prediction. Note that xpred should have the same column names with Xfit_df used in wass_regress_res}

\item{level}{confidence level}

\item{delta}{boundary control value in density band computation. Must be a value in the interval (0, 1/2); (default: 0.01)}

\item{type}{'density', 'quantile' or 'both'}

\item{figure}{logical; if TRUE, return a sampled plot (default: TRUE)}

\item{fig_num}{row of \code{Xpred_df} used for computation of confidence bands. If NULL, then \code{fig_num} is randomly chosen}
}
\value{
a list containing the following lists:
\item{den_list:}{
\itemize{
\item{fpred:} {k-by-m' matrix, predicted density function at xpred.}
\item{f_ux:} {k-by-m' matrix, upper bound of confidence bands in terms of density functions.}
\item{f_lx:} {k-by-m' matrix, lower bound of confidence bands in terms of density functions.}
\item{Qpred:} {k-by-m'matrix, f_lx[i, ], f_ux[i, ] and fpred[i, ] evaluated on Qpred[i, ] vector.}
}}
\item{quan_list:}{
\itemize{
\item{Qpred:} {p-by-m matrix of predicted quantile functions.}
\item{Q_ux:} {p-by-m matrix of upper bound of quantile functions.}
\item{Q_lx:} {p-by-m matrix of lower bound of quantile functions.}
\item{t:} {a length m vector - common grid for all quantile functions.}}}
\item{cdf_list:}{
\itemize{
\item{fpred:} {p-by-m matrix, predicted density function}
\item{Fpred:} {p-by-m matrix, predicted cumulative distribution functions.}
\item{F_ux:} {p-by-m matrix, upper bound of cumulative distribution functions.}
\item{F_lx:} {p-by-m matrix, lower bound of cumulative distribution functions.}
\item{Fsup:} {p-by-m matrix, F_lx, F_ux and Fpred evaluated on Fsup[i, ] vector.}}}
}
\description{
Confidence Bands for Wasserstein Regression
}
\details{
This function computes intrinsic confidence bands for \code{Xpred_df} if \code{type} = 'quantile' and density bands if \code{type} = 'density'. Plots the confidence and/or density bands when \code{figure} = TRUE.
}
\examples{
data(strokeCTdensity)
predictor = strokeCTdensity$predictors
dSup = strokeCTdensity$densitySupport
densityCurves = strokeCTdensity$densityCurve
xpred = colMeans(predictor)

res = wass_regress(rightside_formula = ~., Xfit_df = predictor, Ymat = densityCurves, bandwidth = 0.05, Ytype = 'density', Sup = dSup)
confidence_Band = confidenceBands(res, Xpred_df = xpred, type = 'density')
}
