% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den2Q_dq.R
\name{den2Q_qd}
\alias{den2Q_qd}
\title{convert density function to quantile and quantile density function}
\usage{
den2Q_qd(densityCurves, dSup, bandwidth = 0.05, t = seq(0, 1, by = 0.01))
}
\arguments{
\item{densityCurves}{n-by-m matrix of density curves}

\item{dSup}{length m vector contains the common support grid of the density curves}

\item{bandwidth}{smoothing parameter used in kernel smooth}

\item{t}{the desired grid points in [0, 1] for the quantile function.}
}
\description{
convert density function to quantile and quantile density function
}
