% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.WRI.R
\name{print.summary.WRI}
\alias{print.summary.WRI}
\title{print the summary of WRI object}
\usage{
\method{print}{summary.WRI}(summaryWRI)
}
\arguments{
\item{summaryWRI}{a 'summary.WRI' object}

\item{digits}{integer indicating the number of decimal places in \code{round} function. (default: 3)}
}
\description{
print the summary of WRI object
}
