% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strokeCTdensity.R
\docType{data}
\name{strokeCTdensity}
\alias{strokeCTdensity}
\title{Stroke data: clinical, radiological scalar variables and density curves of the hematoma of 393 stroke patients}
\format{a list of the following three feilds:
\describe{
\item{densityCurve:}{393-by-101 head CT hemotoma densities as distributional response}
\item{densitySupport:}{length 101 common support vector}
\item{predictors:}{393-by-9 matrix containing 9 scalar predictors}
}}
\description{
Stroke data: clinical, radiological scalar variables and density curves of the hematoma of 393 stroke patients
}
\references{
We thank Dr. Divani for providing the original data.
}
