% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.WRI.R
\name{summary.WRI}
\alias{summary.WRI}
\title{Summary Function of Wasserstein Regression Model}
\usage{
\method{summary}{WRI}(WRIobject, digits = 3)
}
\arguments{
\item{wass_regress_res}{an object returned by the \code{wass_regress} function}
}
\value{
a list containing the following fields:
\item{call}{function call of the Wasserstein regression}
\item{method}{methods used to compute p value, see details}
\item{statistic}{test statistics}
\item{critical_value}{critical value}
\item{p_value}{p value of global F test}
\item{wasserstein.F_stat}{Wasserstein F statistic value from the Satterthwaite method}
\item{chisq_df}{degrees of freedom of the null chi-square distribution}
}
\description{
Summary Function of Wasserstein Regression Model
}
\examples{
data(strokeCTdensity)
predictor = strokeCTdensity$predictors
dSup = strokeCTdensity$densitySupport
densityCurves = strokeCTdensity$densityCurve

res <- wass_regress(rightside_formula = ~., Xfit_df = predictor, Ymat = densityCurves, Ytype = 'density', Sup = dSup)
summary(res)
}
